Renamed sw.py to splitwise.py
